backend/requirements.txt
fastapi
uvicorn[standard]
pandas
yfinance
numpy
python-dotenv
psycopg2-binary 